order alerts

+++ app/views/productionmanager/outletorders.php
@@ -36,6 +36,10 @@
         .blue {
             background-color: #3498db;
         }
+        .yellow{
+            background-color: #f1c40f;
+            margin-left: 10px;
+        }
     </style>
     <title>Outlet Orders</title>
 </head>
@@ -89,7 +93,7 @@
             </tr>";
             
             foreach ($productorder as $ProductOrder){
-                if ($ProductOrder->orderstatus == "pending"){
+                if ($ProductOrder->orderstatus == "pending" && $ProductOrder->orderdate >= date('Y-m-d')){
                     echo "<tr>";
                     echo "<td>".$ProductOrder->orderref."</td>";
                     echo "<td>".$ProductOrder->placeby."</td>";
@@ -100,12 +104,7 @@
                     echo "<td>".$ProductOrder->deliverby."</td>";
                     echo "<td>".$ProductOrder->unique_id."</td>";
                     echo "<td>".$ProductOrder->deliver_address."</td>";
-                    if($ProductOrder->paymentstatus == "paid" || $ProductOrder->paymentstatus == "advanced"){
-                        echo "<td><button class='button green' onclick='process(".$ProductOrder->orderid.", \"".$ProductOrder->deliverystatus."\")'>Process</button></td>";
-                    }
-                    else{
-                        echo "<td>Incomplete Payment</td>";
-                    }
+                    echo "<td><button class='button green' onclick='process(".$ProductOrder->orderid.", \"".$ProductOrder->deliverystatus."\")'>Process</button></td>";
                     echo "<td><button class='button red' onclick='cancel(".$ProductOrder->orderid.")'>Cancel</button></td>";
                     echo "<td><button class='button blue' onclick='more(\"" . $ProductOrder->unique_id . "\")'>More</button></td>";
                     echo "</tr>";
@@ -135,7 +134,7 @@
                 </tr>";
                 
                 foreach ($productorder as $ProductOrder){
-                    if ($ProductOrder->orderstatus == "processing" && ($ProductOrder->paymentstatus == "paid" || $ProductOrder->paymentstatus == "advanced") ){ //&& $ProductOrder->orderdate == date('Y-m-d', strtotime('+1 day'))
+                    if ($ProductOrder->orderstatus == "processing" && ($ProductOrder->paymentstatus == "paid" || $ProductOrder->paymentstatus == "advanced") && $ProductOrder->orderdate == date('Y-m-d', strtotime('+1 day'))){ 
                         echo "<tr>";
                         echo "<td>".$ProductOrder->orderref."</td>";
                         echo "<td>".$ProductOrder->placeby."</td>";
@@ -146,7 +145,7 @@
                         echo "<td>".$ProductOrder->deliverby."</td>";
                         echo "<td>".$ProductOrder->unique_id."</td>";
                         echo "<td>".$ProductOrder->deliver_address."</td>";
-                        echo "<td><button class='button green' onclick='completeProduction(".$ProductOrder->orderid.", \"".$ProductOrder->deliverystatus."\")'>Complete Production</button></td>";
+                        echo "<td><button class='button green' onclick='completeProduction(".$ProductOrder->orderid.", \"".$ProductOrder->deliverystatus."\")'>Production Complete</button></td>";
                         echo "<td><button class='button red' onclick='cancel(".$ProductOrder->orderid.")'>Cancel</button></td>";
                         echo "<td><button class='button blue' onclick='more(\"" . $ProductOrder->unique_id . "\")'>More</button></td>";
                         echo "</tr>";
@@ -273,8 +272,34 @@
     }
 
     function process(orderid){
-        var url = BASE_URL + "pmcontrols/processOrderOutlet/" + orderid;
-        window.location.href = url;
+        const SwalwithButton = Swal.mixin({
+            customClass: {
+                confirmButton: "button yellow",
+                cancelButton: "button yellow"
+            },
+            buttonsStyling: false
+            });
+
+            SwalwithButton.fire({
+                text: "Are you sure you want to process this order?",
+                icon: "warning",
+                showCancelButton: true,
+                confirmButtonText: "Yes",
+                cancelButtonText: "No",
+            }).then((result) => {
+                if (result.isConfirmed) {
+                    Swal.fire({
+                        icon: 'success',
+                        title: 'Order Processed Successfully',
+                        text: 'The order has been successfully processed.',
+                        confirmButtonText: 'OK'
+                    }).then(() => {
+                        var url = BASE_URL + "pmcontrols/processOrderOutlet/" + orderid;
+                        window.location.href = url;
+                    });
+                }
+            });
+       
     }
 
     function completeProduction(orderid, deliverystatus) {
@@ -289,8 +314,33 @@
     }
 
     function cancel(orderid) {
-        var url = BASE_URL + "pmcontrols/cancelOrderOutlet/" + orderid;
-        window.location.href = url;
+        const SwalwithButton = Swal.mixin({
+            customClass: {
+                confirmButton: "button yellow",
+                cancelButton: "button yellow"
+            },
+            buttonsStyling: false
+            });
+
+            SwalwithButton.fire({
+                text: "Are you sure you want to cancel this order?",
+                icon: "warning",
+                showCancelButton: true,
+                confirmButtonText: "Yes",
+                cancelButtonText: "No",
+            }).then((result) => {
+                if (result.isConfirmed) {
+                    Swal.fire({
+                        icon: 'success',
+                        title: 'Order Canceled Successfully',
+                        text: 'The order has been successfully canceled.',
+                        confirmButtonText: 'OK'
+                    }).then(() => {
+                        var url = BASE_URL + "pmcontrols/cancelOrderOutlet/" + orderid;
+                        window.location.href = url;
+                    });
+                }
+            });
     }
 
     function more(unique_id) {
@@ -305,9 +355,34 @@
     }
 
     function completed(orderid){
-        sessionStorage.setItem('activeLink', 'showCompletedOrderTable(this)')
-        var url = BASE_URL + "pmcontrols/completeOrderOutlet/" + orderid;
-        window.location.href = url;
+        const SwalwithButton = Swal.mixin({
+            customClass: {
+                confirmButton: "button yellow",
+                cancelButton: "button yellow"
+            },
+            buttonsStyling: false
+            });
+
+            SwalwithButton.fire({
+                text: "Are you sure you want to complete this order?",
+                icon: "warning",
+                showCancelButton: true,
+                confirmButtonText: "Yes",
+                cancelButtonText: "No",
+            }).then((result) => {
+                if (result.isConfirmed) {
+                    Swal.fire({
+                        icon: 'success',
+                        title: 'Order Completed Successfully',
+                        text: 'The order has been successfully Completed.',
+                        confirmButtonText: 'OK'
+                    }).then(() => {
+                        sessionStorage.setItem('activeLink', 'showCompletedOrderTable(this)')
+                        var url = BASE_URL + "pmcontrols/completeOrderOutlet/" + orderid;
+                        window.location.href = url;
+                    });
+                }
+            });  
     }
 
     function showPendingOrderTable(link){
